refactored header handling and improved protocol header implementations
