+++ header/header.go
@@ -1,54 +1,54 @@
-package protocol
+package header
 
 import (
 	"github.com/jexia/maestro/refs"
 	"github.com/jexia/maestro/specs"
 )
 
-// A Header represents the key-value pairs.
-type Header map[string]string
+// Store represents the key-value pairs.
+type Store map[string]string
 
 // Clone returns a copy of h or nil if h is nil.
-func (h Header) Clone() Header {
+func (h Store) Clone() Store {
 	return h
 }
 
 // Del deletes the values associated with key.
-func (h Header) Del(key string) {
+func (h Store) Del(key string) {
 	delete(h, key)
 }
 
 // Get gets the first value associated with the given key. If there are no values associated with the key, Get returns "".
-func (h Header) Get(key string) string {
+func (h Store) Get(key string) string {
 	return h[key]
 }
 
 // Set sets the header entries associated with key to the single element value. It replaces any existing values associated with key.
-func (h Header) Set(key, value string) {
+func (h Store) Set(key, value string) {
 	h[key] = value
 }
 
-// NewHeaderManager constructs a new header manager for the given resource
-func NewHeaderManager(resource string, params *specs.ParameterMap) *HeaderManager {
-	return &HeaderManager{
+// NewManager constructs a new header manager for the given resource
+func NewManager(resource string, params *specs.ParameterMap) *Manager {
+	return &Manager{
 		Resource: specs.JoinPath(resource, specs.ResourceHeader),
 		Params:   params,
 	}
 }
 
-// HeaderManager represents a header manager for a given resource
-type HeaderManager struct {
+// Manager represents a header manager for a given resource
+type Manager struct {
 	Resource string
 	Params   *specs.ParameterMap
 }
 
 // Marshal attempts to marshal the given header specs from the given refs store
-func (manager *HeaderManager) Marshal(store *refs.Store) Header {
+func (manager *Manager) Marshal(store *refs.Store) Store {
 	if manager == nil || manager.Params == nil {
-		return make(Header, 0)
+		return make(Store, 0)
 	}
 
-	result := make(Header, len(manager.Params.Header))
+	result := make(Store, len(manager.Params.Header))
 	for key, property := range manager.Params.Header {
 		value := property.Default
 
@@ -70,7 +70,7 @@ func (manager *HeaderManager) Marshal(store *refs.Store) Header {
 }
 
 // Unmarshal unmarshals the given protocol header into the given reference store
-func (manager *HeaderManager) Unmarshal(header Header, store *refs.Store) {
+func (manager *Manager) Unmarshal(header Store, store *refs.Store) {
 	for key, value := range header {
 		ref := refs.New(key)
 		ref.Value = value
